feat: New layer options
Now you can rename a layer by double clicking on its name.
Additionally, you can change the position of the layers using two icons that are shown in the layers section.

Slight improvement in the keyframe creation function (handleNewEmptyFrame):
Now the function needs a method as a parameter, which defines whether the new keyframes will be empty or a copy of the last keyframe of the layer.