feat: break animation and loop mode
The animation can now be stopped thanks to a new icon implemented in the timeline header.
Furthermore, the animation is no longer played by clicking on a keyframe,
but from the same icon that is modified depending on the "isPlaying" state.
The animation breaking method is managed thanks to the use of a ref called "shouldStopRef".

In addition, there is now a new playback mode: "loop mode".
It is established from the header of the timeline.